backend/pyproject.toml
[tool.poetry]
name = "ssdl-backend"
version = "0.1.0"
description = "Backend for SSDL Calibration Web Application"
authors = ["SSDL Lab <lab@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
pydantic = "^2.9.2"
pint = "^0.24.3"
pandas = "^2.2.3"
numpy = "^1.26.4"
python-docx = "^1.1.2"
whoosh = "^2.7.4"
alembic = "^1.13.3"
fastapi-users = {extras = ["sqlalchemy"], version = "^14.0.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pytest = "^8.3.3"

[tool.poetry.dev-dependencies]
black = "^24.8.0"
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
alembic = "^1.15.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"